office changes - clips and some conversions
added and resized a lot of clips, opened the elevator and worked on T-wing to remove an unnatural player clip.  added stair/sidewalk clips for zspawn on one side and changed some models to prop_physics.  made the vending machine in the kitchen physics and deleted the lighting entities, doesn't make sense to raise the file size of the map for something so small